fix(edgeless): when element resize, should check connector
