+++ packages/blocks/src/page-block/edgeless/mode-controllers/default-mode.ts
@@ -15,8 +15,7 @@ import {
   type TopLevelBlockModel,
 } from '@blocksuite/blocks/std';
 import { assertExists, caretRangeFromPoint } from '@blocksuite/global/utils';
-import type { ConnectorElement, PhasorElement, XYWH } from '@blocksuite/phasor';
-import { getBrushBoundFromPoints } from '@blocksuite/phasor';
+import type { PhasorElement, XYWH } from '@blocksuite/phasor';
 import { deserializeXYWH, getCommonBound, isPointIn } from '@blocksuite/phasor';
 
 import { showFormatQuickBar } from '../../../components/format-quick-bar/index.js';
@@ -28,9 +27,9 @@ import {
 } from '../../utils/position.js';
 import type { Selectable } from '../selection-manager.js';
 import {
-  generateConnectorPath,
-  getConnectorAttachedInfo,
   getXYWH,
+  handleElementChangedEffectForConnector,
+  isConnectorAndBindingsAllSelected,
   isPhasorElement,
   isTopLevelBlock,
   pickBlocksByBound,
@@ -47,32 +46,6 @@ enum DragType {
   NativeEditing = 'native-editing',
   /** Default void state */
   None = 'none',
-}
-
-function isConnectorAndBindingsAllSelected(
-  connector: ConnectorElement,
-  selected: Selectable[]
-) {
-  const connectorSelected = selected.find(s => s.id === connector.id);
-  if (!connectorSelected) {
-    return false;
-  }
-  const { startElement, endElement } = connector;
-  const startSelected = selected.find(s => s.id === startElement?.id);
-  const endSelected = selected.find(s => s.id === endElement?.id);
-  if (!startElement && !endElement) {
-    return true;
-  }
-  if (!startElement && endSelected) {
-    return true;
-  }
-  if (!endElement && startSelected) {
-    return true;
-  }
-  if (startSelected && endSelected) {
-    return true;
-  }
-  return false;
 }
 
 export class DefaultModeController extends MouseModeController<DefaultMouseMode> {
@@ -148,6 +121,15 @@ export class DefaultModeController extends MouseModeController<DefaultMouseMode>
     }
   }
 
+  private _handleDragMoveEffect(element: Selectable) {
+    handleElementChangedEffectForConnector(
+      element,
+      this._blockSelectionState.selected,
+      this._edgeless.surface,
+      this._page
+    );
+  }
+
   private _handleSurfaceDragMove(selected: PhasorElement, e: SelectionEvent) {
     if (!this._lock) {
       this._lock = true;
@@ -178,57 +160,7 @@ export class DefaultModeController extends MouseModeController<DefaultMouseMode>
       });
     }
 
-    if (selected.type !== 'connector') {
-      const bindingElements = surface.getBindingElements(selected.id);
-      bindingElements.forEach(bindingElement => {
-        if (bindingElement.type === 'connector') {
-          if (
-            isConnectorAndBindingsAllSelected(
-              bindingElement,
-              this._blockSelectionState.selected
-            )
-          ) {
-            return;
-          }
-          const { startElement, endElement, id, x, y, controllers, mode } =
-            bindingElement;
-          const { start, end } = getConnectorAttachedInfo(
-            bindingElement,
-            surface,
-            this._page
-          );
-          const fixed =
-            startElement?.id === selected.id
-              ? 'end'
-              : endElement?.id === selected.id
-              ? 'start'
-              : undefined;
-
-          const routes = generateConnectorPath(
-            start.rect,
-            end.rect,
-            start.point,
-            end.point,
-            controllers.map(c => ({ ...c, x: c.x + x, y: c.y + y })),
-            mode,
-            fixed
-          );
-
-          const bound = getBrushBoundFromPoints(
-            routes.map(r => [r.x, r.y]),
-            0
-          );
-          const newControllers = routes.map(v => {
-            return {
-              ...v,
-              x: v.x - bound.x,
-              y: v.y - bound.y,
-            };
-          });
-          surface.updateConnectorElement(id, bound, newControllers);
-        }
-      });
-    }
+    this._handleDragMoveEffect(selected);
   }
 
   private _handleBlockDragMove(block: TopLevelBlockModel, e: SelectionEvent) {
@@ -241,6 +173,7 @@ export class DefaultModeController extends MouseModeController<DefaultMouseMode>
       modelH,
     ]);
     this._page.updateBlock(block, { xywh });
+    this._handleDragMoveEffect(block);
   }
 
   private _isInSelectedRect(viewX: number, viewY: number) {
